Move shared parametrized_api_urls to util file
